// angular-sample/angular-webapp/client/app/company/manage-departments/models/employee-basic-info.model.ts
import { Employee, EmployeeContacts } from '../../../employee/models/employee.model';
import * as Immutable from 'immutable';
import { EmergencyContact } from '../../../employee/models/emergency-contact.model';

export class EmployeeBasicInfoModel {
    Id: string;
    Title: string;
    Salutation: string;
    FullName: string;
    FirstName: string;
    MiddleName: string;
    Surname: string;
    KnownAs: string;
    PreviousName: string;
    Gender: number;
    GenderText: string;
    DOB?: Date;
    EthnicGroupValueName: string;
    EthnicGroupName: string;
    EthnicGroupValueId: string;
    EthnicGroupValueType: number;
    Nationality: string;
    Age: string;
    TaxCode: string;
    NINumber: string;
    UserId: string;
    CompanyId: string;
    PersonalEmail: string;
    Email: string;
    AddressLine1: string;
    AddressLine2: string;
    AddressLine3: string;
    Town: string;
    CountyId: string;
    CountyName: string;
    CountryId: string;
    CountryName: string;
    Postcode: string;
    HomePhone: string;
    MobilePhone: string;
    FullAddress: string;

    constructor() {
        this.Id = '';
        this.FirstName = '';
        this.MiddleName = '';
        this.Surname = '';
        this.DOB = new Date();
        this.Title = '';
        this.FullName = '';
        this.Salutation = '';
        this.Age = '';
        this.Nationality = '';
        this.GenderText = '';
        this.Gender = 0;
        this.PreviousName = '';
        this.KnownAs = '';
        this.UserId = '';
        this.EthnicGroupValueName = '';
        this.EthnicGroupName = '';
        this.EthnicGroupValueId = '';
        this.EthnicGroupValueType = null;
        this.CompanyId = '';
        this.TaxCode = '';
        this.NINumber = '';
        this.PersonalEmail = '';
        this.Email = '';
        this.AddressLine1 = '';
        this.AddressLine2 = '';
        this.AddressLine3 = '';
        this.Town = '';
        this.CountyId = '';
        this.CountyName = '';
        this.CountryId = '';
        this.CountryName = '';
        this.Postcode = '';
        this.HomePhone = '';
        this.MobilePhone = '';
        this.FullAddress = '';
    }
}